// frontend/src/api/openapi/models/LLMPromptTemplates.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type LLMPromptTemplates = {
  /**
   * The language of the prompt template
   */
  language: string;
  /**
   * The system prompt to use for the job
   */
  system_prompt: string;
  /**
   * The user prompt to use for the job
   */
  user_prompt: string;
};